refactor: mvdir to alpha-components
