+++ packages/bezier-react/src/alpha-components/CheckboxPrimitive/CheckboxPrimitive.types.ts
