+++ packages/bezier-react/src/components/Modal/Modal.tsx
@@ -27,8 +27,7 @@ import {
   type DialogPrimitiveProps,
   DialogPrimitiveTitle,
   DialogPrimitiveTrigger,
-} from '~/src/primitives/DialogPrimitive'
-
+} from '~/src/alpha-components/DialogPrimitive'
 import { Button } from '~/src/components/Button'
 import { Text } from '~/src/components/Text'
 import {
